+++ HHH_semiboosted_VR.py
@@ -209,7 +209,7 @@ def test_make(jsonConfig,findreplace={}):
     twoD.Save()
     
 
-def test_fit(strategy=0):
+def test_fit(strategy=0, rMin=-1, rMax=10):
     twoD = TwoDAlphabet(working_area, '%s/runConfig.json'%working_area, loadPrevious=True)
     subset = twoD.ledger.select(_select_bkg, polyOrder)
     twoD.MakeCard(subset, '{0}_area'.format(polyOrder))
@@ -222,7 +222,7 @@ def test_fit(strategy=0):
     #     print("Fit cmd: ", fitCmd)
     #     os.system(fitCmd)
 
-    twoD.MLfit('{0}_area'.format(polyOrder),strategy=strategy,verbosity=0)
+    twoD.MLfit('{0}_area'.format(polyOrder),strategy=strategy,rMin=rMin,rMax=rMax,verbosity=0)
 
 def test_limit(working_area,orderSR,json_file,blind=True):
     '''Perform a blinded limit. To be blinded, the Combine algorithm (via option `--run blind`)
@@ -412,22 +412,22 @@ if __name__ == '__main__':
 
     (options, args) = parser.parse_known_args()
 
-    bestOrder = {"{}_semiboosted_CR_pass_toy".format(options.year):"1"}
-    for working_area in ["{}_semiboosted_CR_pass_toy".format(options.year)]:
+    bestOrder = {"{}_semiboosted_VR".format(options.year):"1"}
+    for working_area in ["{}_semiboosted_VR".format(options.year)]:
 
         jsonConfig   = 'configs/HHH/{0}.json'.format(working_area)
 
         test_make(jsonConfig)
 
-        for order in ["0","1","2"]:
+        for order in ["0","1","2","3"]:
             polyOrder = order
             if options.year == "2017":
-                if polyOrder in ["2"]:
+                if polyOrder in ["3"]:
                     test_fit(strategy=1)
                 else:
                     test_fit()
             else:
-                test_fit()
+                test_fit(strategy=1, rMin=-5, rMax=5)
             test_plot()
             if polyOrder==bestOrder[working_area]:
                 test_GoF() # this waits for toy fits on Condor to finish
@@ -435,4 +435,4 @@ if __name__ == '__main__':
 
         test_FTest("0","1")
         test_FTest("1","2")
-        #test_FTest("2","3")
+        test_FTest("2","3")
