+++ README.md
@@ -40,24 +40,24 @@ source activate_env
 
 For running fits and making plots for the boosted control region, run
 ```
-python -u HHH_2017_boosted_CR.py |& tee 2017_boosted_CR_`date "+%Y%m%d_%H%M%S"`.log
+python -u HHH_2017_boosted_VR.py |& tee 2017_boosted_VR_`date "+%Y%m%d_%H%M%S"`.log
 ```
 To do the same for the semiboosted control region, run
 ```
-python -u HHH_2017_semiboosted_CR.py |& tee 2017_semiboosted_CR_`date "+%Y%m%d_%H%M%S"`.log
+python -u HHH_2017_semiboosted_VR.py |& tee 2017_semiboosted_VR_`date "+%Y%m%d_%H%M%S"`.log
 ```
 Note that piping output to the `tee` command will both print it to the terminal and save it in a log file. The log file name will contain a timestamp.
 
 To calculate expected limits, we first need to generate toy data in the pass category of the signal regions. For this we use the pass-to-fail transfer functions (Rpf) obtained from the control region fits. First we need to extract the fit parameter values which can be done using the following command
 ```
-echo -e "Boosted CR:\nOrder 1 (best)" |& tee printFitParameters_CR_`date "+%Y%m%d"`.log
-python printFitParameters.py -i 2017_boosted_CR/1_area/fitDiagnosticsTest.root |& tee -a printFitParameters_CR_`date "+%Y%m%d"`.log
-echo -e "\n\nSemiboosted CR:\nOrder 1 (best)" |& tee -a printFitParameters_CR_`date "+%Y%m%d"`.log
-python printFitParameters.py -i 2017_semiboosted_CR/1_area/fitDiagnosticsTest.root |& tee -a printFitParameters_CR_`date "+%Y%m%d"`.log
+echo -e "Boosted VR:\nOrder 1 (best)" |& tee printFitParameters_VR_`date "+%Y%m%d"`.log
+python printFitParameters.py -i 2017_boosted_VR/1_area/fitDiagnosticsTest.root |& tee -a printFitParameters_VR_`date "+%Y%m%d"`.log
+echo -e "\n\nSemiboosted VR:\nOrder 1 (best)" |& tee -a printFitParameters_VR_`date "+%Y%m%d"`.log
+python printFitParameters.py -i 2017_semiboosted_VR/1_area/fitDiagnosticsTest.root |& tee -a printFitParameters_VR_`date "+%Y%m%d"`.log
 ```
 The printed fit parameter values need to be put into `plotRpf.py` in oder to plot the transfer functions
 ```
-python plotRpf.py
+python plotRpf_VR.py
 ```
 The toy data is generated using the `generateToys.py` script which imports the transfer functions from `plotRpf.py`
 ```
@@ -67,21 +67,21 @@ Two output files are produced, `JetHT_Histograms_VR_pass_toy.root` with the toy 
 ```
 mv JetHT_Histograms_*_pass_toy.root symlink2histograms_2017
 ```
-The toy data in the control regions is used in `HHH_2017_boosted_CR_pass_toy.py` and `HHH_2017_semiboosted_CR_pass_toy.py` as a sort of sanity check (closure test) to check whether the toy data fits converge to parameter values similar to those used in the generation of the toy data
+The toy data in the control regions is used in `HHH_2017_boosted_VR_pass_toy.py` and `HHH_2017_semiboosted_VR_pass_toy.py` as a sort of sanity check (closure test) to check whether the toy data fits converge to parameter values similar to those used in the generation of the toy data
 ```
-python -u HHH_2017_boosted_CR_pass_toy.py |& tee 2017_boosted_CR_pass_toy_`date "+%Y%m%d_%H%M%S"`.log
-python -u HHH_2017_semiboosted_CR_pass_toy.py |& tee 2017_semiboosted_CR_pass_toy_`date "+%Y%m%d_%H%M%S"`.log
+python -u HHH_2017_boosted_VR_pass_toy.py |& tee 2017_boosted_VR_pass_toy_`date "+%Y%m%d_%H%M%S"`.log
+python -u HHH_2017_semiboosted_VR_pass_toy.py |& tee 2017_semiboosted_VR_pass_toy_`date "+%Y%m%d_%H%M%S"`.log
 ```
 The Rpf parameter values can be printed with the following commands
 ```
-echo -e "Boosted CR pass toy:\nOrder 1 (best)" |& tee printFitParameters_CR_pass_toy_`date "+%Y%m%d"`.log
-python printFitParameters.py -i 2017_boosted_CR_pass_toy/1_area/fitDiagnosticsTest.root |& tee -a printFitParameters_CR_pass_toy_`date "+%Y%m%d"`.log
-echo -e "\n\nSemiboosted CR pass toy:\nOrder 1 (best)" |& tee -a printFitParameters_CR_pass_toy_`date "+%Y%m%d"`.log
-python printFitParameters.py -i 2017_semiboosted_CR_pass_toy/1_area/fitDiagnosticsTest.root |& tee -a printFitParameters_CR_pass_toy_`date "+%Y%m%d"`.log
+echo -e "Boosted VR pass toy:\nOrder 1 (best)" |& tee printFitParameters_VR_pass_toy_`date "+%Y%m%d"`.log
+python printFitParameters.py -i 2017_boosted_VR_pass_toy/1_area/fitDiagnosticsTest.root |& tee -a printFitParameters_VR_pass_toy_`date "+%Y%m%d"`.log
+echo -e "\n\nSemiboosted VR pass toy:\nOrder 1 (best)" |& tee -a printFitParameters_VR_pass_toy_`date "+%Y%m%d"`.log
+python printFitParameters.py -i 2017_semiboosted_VR_pass_toy/1_area/fitDiagnosticsTest.root |& tee -a printFitParameters_VR_pass_toy_`date "+%Y%m%d"`.log
 ```
-The printed fit parameter values need to be put into `plotRpf_CR_pass_toy.py` in order to plot the transfer functions
+The printed fit parameter values need to be put into `plotRpf_VR_pass_toy.py` in order to plot the transfer functions
 ```
-python plotRpf_CR_pass_toy.py
+python plotRpf_VR_pass_toy.py
 ```
 
 To perform the same checks for the signal region fits and calculate expected limits for the benchmark XToYHTo6B_MX-2500_MY-800 case for the boosted channel, run
